add track release
