+++ lib/audioreleaser.rb
@@ -2,8 +2,9 @@
 
 require 'audioreleaser/encoder'
 require 'audioreleaser/album'
+require 'audioreleaser/album_release'
 require 'audioreleaser/track'
+require 'audioreleaser/track_release'
 require 'audioreleaser/license'
 require 'audioreleaser/tagger'
 require 'audioreleaser/tag'
-require 'audioreleaser/release'
